A number of edits and changes. Below is a reproduction of the comments I'll send separately.
A few notes/observations in no particular order:•	I made a few edits to clarify variable naming and streamline a few calculations (while trying to stay away from big changes to the code that would confuse you). Let me know if anything is perplexing.•	I also did some basic checks to make sure that the code still works. Hopefully I didn’t mess anything up with these changes.•	The period indexing is really hard to follow, with optInputs.PerTgt, perKnown, recPer and perKnownCorr being subtly different variables that are hard to differentiate from the variable names, and there are seemingly repeated lines where the conditioning period is being added and removed. I know that this complexity snuck up on us, so it's understandable, but we should clean it up before publishing. I have a few thoughts:o	From a blank slate, what would you say are the key variables related to periods that are needed in the code? I can think of:♣	All periods of interest for the spectrum calculations and record searching (I think this is    optInputs.PerTgt)♣	The index of the conditioning period, if there is one (optInputs.rec)♣	The indices of all other periods (I think this is notT1)♣	Perhaps a set of indices to map the periods of interest back to the original record spectra matrix?♣	Anything else?o	If there is a way to simplify the code down to the above variables, great. If not, can you describe the definitions of the above variables, and then we can work on more descriptive variables names so that it's clear what their differences are?•	Is there a reason for having finalScaleFactors and finalScaleFac be two distinct variables? It looks like they are the same thing and that you switch names partway through the code (finalScaleFactors = finalScaleFac;)•	I took out a couple of plots in plotResults that I thought were redundant with the others, just to simplify things•	In the ComputeTargets function, I have a few questionso	You seem to use variable names for covariances and correlations interchangably, which makes it very hard to follow the calculation (e.g., Targets.covReq = corrMatrix). Perhaps some of that is old legacy variable names, please recheck it and make the variable names refer to the proper propertyo	Why the switch in variable names from covReqPart to covReqPartNew at line 90?o	Same question with corrReqNew and corrMatrix at line 88o	Add a short comment at line 89 and 91, so the reader knows what's going on here •	I tried to reorganize and shorten the variable documentation section. A few notes on thiso	In general, I took out any commentary ("This variable will be used to...") as we really just want to define the variables here and not discuss the algorithm. The user can read the later code or your paper to learn more about the algorithm.o	Where it seemed feasible, I put the variable definition inline with where the variable is actually assigned, so that a user can see the definition quickly. Probably a few more could be rearranged like this—see what you think.o	In a few of these definitions, you use the phrase “allowed spectra”—is this the set of spectra associated with ground motions that are being searched? I think I would make this more explicit, as it’s not clear at first reading what “allowed” means o	It’s good that you defined some other variables being used (SaKnown, soil_Vs30, etc.), but let’s move those to a location separate from where the user inputs are defined. Otherwise when I read the comments it gives me the impression that I will need to define these at the next step. Perhaps move those comments to after the user inputs, or define them inline in the code when they are first used?•	I re-wrote the initial search for ground motions (prior to the optimization loop), and I think it’s a bit cleaner now. Take a look and let me know if you see a problem with it that I missed.•	Is the KS_stat function still used? I didn’t see it in use, and it could be removed if so.•	The extra line with the SB_2014_ratios function is to adjust our target in the case where we are using a RotD100 target (in that case the mean of the spectrum is larger, and the standard deviation is slightly larger). I’ve been meaning to address that for a while and kept forgetting. It doesn’t affect the rest of the code, but I wanted to let you know why that’s in there now.